chore: Avoid implicit cast in math operations
Resolve Sonar issues of type java:S2184 - Math operands should be cast before assignment

related to #88

+++ engine-dmn/engine/src/main/java/org/operaton/bpm/dmn/engine/impl/evaluation/DecisionTableEvaluationHandler.java
@@ -95,7 +95,7 @@ public class DecisionTableEvaluationHandler implements DmnDecisionLogicEvaluatio
   }
 
   protected long calculateExecutedDecisionElements(DmnDecisionTableImpl decisionTable) {
-    return (decisionTable.getInputs().size() + decisionTable.getOutputs().size()) * decisionTable.getRules().size();
+    return (decisionTable.getInputs().size() + decisionTable.getOutputs().size()) * (long) decisionTable.getRules().size();
   }
 
   protected void evaluateDecisionTable(DmnDecisionTableImpl decisionTable, VariableContext variableContext, DmnDecisionTableEvaluationEventImpl evaluationResult) {
